+++ mobtravelapp/src/main/java/com/mobiquity/mobtravelapp/controller/TravelController.java
@@ -1,7 +1,8 @@
 package com.mobiquity.mobtravelapp.controller;
 
-import com.mobiquity.mobtravelapp.model.RouteModel;
-import com.mobiquity.mobtravelapp.model.Trip;
+import com.mobiquity.mobtravelapp.model.travelModel.Leg;
+import com.mobiquity.mobtravelapp.model.travelModel.RouteModel;
+import com.mobiquity.mobtravelapp.model.travelModel.Trip;
 import com.mobiquity.mobtravelapp.service.TravelService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -16,8 +17,8 @@ public class TravelController {
     TravelService travelService;
 
     @GetMapping("/routes")
-    public Trip getRoutes(@RequestParam(value ="fromStation") String fromStation, @RequestParam(value = "toStation") String toStation, @RequestParam(value = "dateTime") String dateTime) throws IOException {
-         return travelService.getRoutes(new RouteModel(fromStation, toStation, dateTime));
+    public Trip getRoutes(@RequestParam(value ="fromStation") String fromStation, @RequestParam(value = "toStation") String toStation, @RequestParam(value = "dateTime") String dateTime, @RequestParam(defaultValue = "0") int routeLimit) throws IOException {
+         return travelService.getRoutes(new RouteModel(fromStation, toStation, dateTime,routeLimit));
     }
