+++ mobtravelapp/src/main/java/com/mobiquity/mobtravelapp/model/travelModel/DestinationStub.java
@@ -1,0 +1,10 @@
+package com.mobiquity.mobtravelapp.model.travelModel;
+
+public class DestinationStub {
+    private String plannedArrivalTime;
+    private String actualArrivalTime;
+    private String plannedTrack;
+    private String actualTrack;
+    private Station station;
+
+}
