+++ mobtravelapp/src/main/java/com/mobiquity/mobtravelapp/model/travelModel/OriginStub.java
@@ -1,0 +1,10 @@
+package com.mobiquity.mobtravelapp.model.travelModel;
+
+public class OriginStub {
+    protected String plannedDepartureTime;
+   protected String actualDepartureTime;
+    protected String plannedTrack;
+    protected String actualTrack;
+    protected Station station;
+
+}
