+++ mobtravelapp/src/main/java/com/mobiquity/mobtravelapp/model/travelModel/Station.java
@@ -1,0 +1,17 @@
+package com.mobiquity.mobtravelapp.model.travelModel;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@AllArgsConstructor
+@Getter
+@Setter
+@NoArgsConstructor
+public class Station {
+    private String name;
+   private String latitude;
+   private String longitude;
+
+}
