+++ mobtravelapp/src/main/java/com/mobiquity/mobtravelapp/model/travelModel/StopStub.java
@@ -1,0 +1,7 @@
+package com.mobiquity.mobtravelapp.model.travelModel;
+
+public class StopStub extends OriginStub{
+    private String plannedArrivalTime;
+    private String actualArrivalTime;
+
+}
