+++ mobtravelapp/src/main/java/com/mobiquity/mobtravelapp/model/travelModel/Trip.java
@@ -1,4 +1,4 @@
-package com.mobiquity.mobtravelapp.model;
+package com.mobiquity.mobtravelapp.model.travelModel;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.util.List;
+
 @AllArgsConstructor
 @Getter
 @Setter
