+++ mobtravelapp/src/test/java/com/mobiquity/mobtravelapp/service/TravelServiceTest.java
@@ -4,9 +4,8 @@ package com.mobiquity.mobtravelapp.service;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-import com.mobiquity.mobtravelapp.model.Route;
-import com.mobiquity.mobtravelapp.model.RouteModel;
-import org.json.simple.JSONObject;
+import com.mobiquity.mobtravelapp.model.travelModel.Route;
+import com.mobiquity.mobtravelapp.model.travelModel.RouteModel;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.jupiter.api.DisplayName;
